# add runtime performance counters