--- a/src/RA_Integration.vcxproj
+++ b/src/RA_Integration.vcxproj
@@ -116,6 +116,7 @@
     <ClCompile Include="services\impl\WindowsFileSystem.cpp" />
     <ClCompile Include="services\impl\WindowsHttpRequester.cpp" />
     <ClCompile Include="services\Initialization.cpp" />
+    <ClCompile Include="services\PerformanceCounter.cpp" />
     <ClCompile Include="services\SearchResults.cpp" />
     <ClCompile Include="ui\drawing\gdi\GDIBitmapSurface.cpp" />
     <ClCompile Include="ui\drawing\gdi\GDISurface.cpp" />
@@ -259,6 +260,7 @@
     <ClInclude Include="services\impl\WindowsHttpRequester.hh" />
     <ClInclude Include="services\Initialization.hh" />
     <ClInclude Include="services\IThreadPool.hh" />
+    <ClInclude Include="services\PerformanceCounter.hh" />
     <ClInclude Include="services\ServiceLocator.hh" />
     <ClInclude Include="services\SearchResults.h" />
     <ClInclude Include="services\TextReader.hh" />