--- a/src/services/PerformanceCounter.hh
+++ b/src/services/PerformanceCounter.hh
@@ -12,7 +12,7 @@ enum class PerformanceCheckpoint
     RuntimeEvents,
     OverlayManagerAdvanceFrame,
     MemoryBookmarksDoFrame,
-    MemoryDialogInvalidate,
+    MemoryInspectorDoFrame,
 
     NUM_CHECKPOINTS
 };