When enabled, writes a log message every 60 seconds detailing average time spent in `RA_DoAchievementsFrame`. 

For the emulator to run at 60 frames per second, each frame has to be processed in less than 16ms. Our goal is to never use more than 2ms of that.

I've retrofitted this logic into an 0.77 DLL, and you can see that with all the windows open (memory inspector, memory bookmarks, rich presence monitor, achievement list, achievement editor), the average is very close to that 2ms limit:
```
Lap averages: total: 1.524
 Runtime: 0.005
 Bookmarks: 0.009
 Inspector: 1.509
```
As an average has some variance, there are several frames that do exceed the 2ms limit. For most emulated systems, that has negligible impact, but more demanding systems like PSX and NDS do sometimes experience slowdown with the memory inspector open.

These numbers are also largely dependent on the host machine and whatever secondary processes are running. My machine is apparently significantly more powerful than many of the machines other devs use, so every bit of time we can shed is important. 

Performing the same test (as closely as possible) with the new DLL yields much better results:
```
Lap averages: total: 0.175
 Runtime: 0.006
 Bookmarks: 0.005
 Inspector: 0.163
```
The new DLL spends almost 90% less time in `RA_DoAchievementsFrame`.

I originally added this code to prove that the new dialog was faster, but I was surprised by how much faster it appears to be.

This logging (and the associated tracking) is disabled by default. To enable, uncomment the `#define PERFORMANCE_COUNTERS` line in `PerformanceCounter.hh`.